backend/requirements.txt
Django>=5.1
djangorestframework>=3.15
django-cors-headers>=4.4
python-dotenv>=1.0
dj-database-url>=2.2
openai>=1.40
psycopg2-binary>=2.9

